requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
gunicorn==20.1.0
openai==0.27.8
redis==5.0.1
Werkzeug==2.3.8